requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
sentence-transformers
rank-bm25
google-generativeai
huggingface-hub
rapidfuzz
requests
numpy
pydantic
streamlit
